Created RecyclerViewListMethods.kt to replace functions which would add data to lists for RecyclerViews in the following classes: - HomeFragment.kt - ChatActivity.kt - CommentsFragment.kt - ProfilePostsFragment.kt - ProfileLikesFragment.kt
Fixed a bug with the addToSearches function which would display more than one student in the search results based on the amount of chats the user is apart of.

Fixed a bug in CommentsFragment.kt which would cause the number of comments to not update if more than one comment is sent without refreshing the fragment.

Other general bug fixes.

Known bug:
- The number of comments displayed in HomeFragment.kt will not update unless the RecyclerView is refreshed. This causes the number of comments in CommentsFragment.kt to be false when reopening the fragment after comments were previously sent.